+++ src/main/java/lapr/project/data/ClienteDB.java
@@ -37,7 +37,7 @@ public class ClienteDB extends DataHandler {
      * @param numCC número do cartão de cidadão do cliente
      * @return o novo cliente criado
      */
-    public Cliente novoCliente(int nif, String nome, String email, int numeroSegurancaSocial, int creditos, String morada, int numCC, String password) {
+    public Cliente novoCliente(int nif, String nome, String email, int numeroSegurancaSocial, double creditos, String morada, int numCC, String password) {
         return new Cliente(nif, nome, email, numeroSegurancaSocial, creditos, morada, numCC, password);
     }
